refactor: split agent CLI setting schemas
