+++ src/shared/schemas/index.ts
@@ -1,6 +1,7 @@
 // Layer 1: Base types (no dependencies on other schema files)
 export * from './identifiers';
 export * from './agent';
+export * from './agentCliSettings';
 export * from './fileFields';
 export * from './proposalFields';
 export * from './toolConfig';
